Fix std min

+++ src/main.cpp
@@ -188,7 +188,7 @@ int main (int argc, char **argv)
 				{
 					auto threshold (nano_pow::reverse(nano_pow::bit_difficulty(difficulty)));
 					// Force threads and lookup if not given
-					auto threads_l (threads != 0 ? threads : std::min (2048UL, driver->threads_get()));
+					auto threads_l (threads != 0 ? threads : std::min (static_cast<size_t> (2048), driver->threads_get()));
 					lookup = (parsed.count("lookup") == 1 ? lookup : 32);
 					lookup_entries = 1ULL << lookup;
 					std::cout << "Tuning for difficulty " << difficulty << " starting with " << threads_l << " threads and " << std::to_string(lookup_entries / (1024 * 1024) * 4) << "MB memory " << std::endl
